+++ src/App.jsx
@@ -1,153 +1,15 @@
-import { useState } from 'react';
-import logo from './images/logo-espaco-mulher.png';
+import logo from './images/logo-espaco-mulher.png'
 
-const ids = Array.from({ length: 20 }, () => crypto.randomUUID())
+import { useItems } from './hooks/use-items'
 
-const FormAddItem = ({ onSubmitItem }) => {
-  const [inputValue, setInputValeu] = useState('')
-  const [selectValue, setSelectValue] = useState('1')
-
-  const handleChangeInput = (e) => setInputValeu(e.target.value)
-  const handleChangeSelect = (e) => setSelectValue(e.target.value)
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-
-    onSubmitItem({
-      id: crypto.randomUUID(),
-      quantity: +selectValue,
-      name: inputValue,
-      stored: false
-    })
-
-    setInputValeu('')
-    setSelectValue('1')
-  }
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <select
-        value={selectValue}
-        onChange={handleChangeSelect}
-        className='input'
-      >
-        {ids.map((id, index) => (
-          <option key={id} value={index + 1}>
-            {index + 1}
-          </option>
-        ))}
-      </select>
-      <input
-        placeholder='Adicione itens'
-        className='input'
-        type="text"
-        value={inputValue}
-        onChange={handleChangeInput}
-        autoFocus
-      />
-      <button className='bg-pink-500 text-white p-1 px-3 rounded'>
-        Adicionar
-      </button>
-    </form>
-  )
-}
-
-const ListOfItems = ({ orderBy, items, onClickCheck, onClickDelete }) => {
-  const sortedItems = orderBy === "stored"
-    ? items.filter((item) => item.stored)
-    : orderBy === "alphabetically"
-      ? items.toSorted((a, b) => a.name > b.name ? 1 : b.name > a.name ? -1 : 0)
-      : items
-
-  return (
-    <ul className='flex gap-5 flex-wrap justify-around w-[720px]'>
-      {sortedItems.map((item) => (
-        <li
-          className='min-w-6 px-3 justify-start '
-          key={item.id}
-        >
-          <input
-            type="checkbox"
-            checked={item.stored}
-            onChange={() => onClickCheck(item.id)}
-            name="item"
-            className='mx-2'
-          />
-          <span className={item.stored ? 'line-through' : ''}>
-            {item.quantity} {item.name}
-          </span>
-          <button onClick={() => onClickDelete(item.id)} >
-            ❌
-          </button>
-        </li>
-      ))
-      }
-    </ul>
-  )
-}
-
-const Filters = ({ orderBy, onChangeOrder }) => (
-  <select name="order-select" className='input' value={orderBy} onChange={onChangeOrder}>
-    <option value="newest">ordenar por mais Recentes</option>
-    <option value="stored">Mostrar guardados</option>
-    <option value="alphabetically">Ordem alfabetica</option>
-  </select>
-)
-
-const Stats = ({ items }) => {
-  let storedItems = items.reduce((acc, item) => item.stored ? acc + 1 : acc, 0)
-  let storedPercentage = items.length === 0 ? 0 : ((storedItems / items.length) * 100).toFixed(0)
-  const singularPlural = items.length === 1 ? "item" : "itens"
-
-  return (
-    <h3 className=' text-white pb-3'>
-      {<p>
-        {`Você tem ${items.length} ${singularPlural} na lista`}
-        {items.length > 0 && <span> e ja guardou {storedItems} ({storedPercentage}%)</span>}
-      </p>
-      }
-    </h3>
-  )
-}
-
-const useItems = () => {
-  const [items, setItems] = useState([])
-  const [orderBy, setOrderBy] = useState('newest')
-
-  const handleSubmitForm = (newItem) => setItems((prev) => [...prev, newItem])
-  const handleClickClearBtn = () => setItems([])
-  const handleChangeOrder = (e) => setOrderBy(e.target.value)
-
-  const handleClickDelete = (id) =>
-    setItems((prev) => prev.filter((item) => item.id != id))
-
-  const handleClickCheck = (id) =>
-    setItems((prev) =>
-      prev.map((item) =>
-        item.id === id ? { ...item, stored: !item.stored } : item,
-      ),
-    )
-  return {
-    items,
-    orderBy,
-    handleSubmitForm,
-    handleChangeOrder,
-    handleClickCheck,
-    handleClickDelete,
-    handleClickClearBtn
-  }
-}
+import { FormAddItem } from './components/form-add-item'
+import { ListOfItems } from './components/list-of-items'
+import { Filters } from './components/filters'
+import { Stats } from './components/stats'
 
 const App = () => {
-  const {
-    items,
-    orderBy,
-    handleSubmitForm,
-    handleChangeOrder,
-    handleClickCheck,
-    handleClickDelete,
-    handleClickClearBtn
-  } = useItems()
+  const state = useItems()
+
   return (
     <>
       <header
@@ -158,26 +20,26 @@ const App = () => {
       <main>
         <section className='bg-blue-900 flex items-center justify-center py-2'>
           <p className='text-white pr-2'>o que você precisa guardar?</p>
-          <FormAddItem onSubmitItem={handleSubmitForm} />
+          <FormAddItem onSubmitItem={state.handleSubmitForm} />
         </section>
       </main>
       <div className='flex flex-col h-[694px] justify-around'>
         <section className='flex flex-1 bg-orange-200 justify-center'>
           <ListOfItems
-            orderBy={orderBy}
-            items={items}
-            onClickCheck={handleClickCheck}
-            onClickDelete={handleClickDelete}
+            orderBy={state.orderBy}
+            items={state.items}
+            onClickCheck={state.handleClickCheck}
+            onClickDelete={state.handleClickDelete}
           />
         </section>
         <footer className=' bg-blue-900 text-center m-0 py-4'>
           <Filters
-            orderBy={orderBy}
-            onChangeOrder={handleChangeOrder}
+            orderBy={state.orderBy}
+            onChangeOrder={state.handleChangeOrder}
           />
-          <Stats items={items} />
+          <Stats items={state.items} />
           <button
-            onClick={handleClickClearBtn}
+            onClick={state.handleClickClearBtn}
             className='text-white bg-orange-600 p-1 rounded '
           >Limpar lista
           </button>
